my_chatbot/requirements.txt
fastapi
uvicorn
transformers
pandas
python-docx
python-pptx
PyPDF2